Bug fixes to make new bot functional.
Changes were made in RobotMap so that ports corresponded with ports on the
new bot, and the nonexistent shooterArmEncoder was set to null to mitigate
runtime errors.
We added commands to the manipulator joystick in OI so that not all the
commands were on the driver joystick.
We added the SpikePuncher subsystem for the spike puncher and its related
commands.
We changed Robot in that we removed the lines in allPeriodic that pushed
data to SmartDashboard, along with adding Robot.spikePuncher and adding a
new import to simplify using AutonomousCommand.
AutonomousCommand was changed to test all the drivetrain motors.
JoystickDrive was changed to remove magic numbers.
We also updated FlywheelOut to go both directions and renamed it
JoystickFlywheel (to indicate that it reads directly from the joystick).