Create separate ramparts autonomous program.
Because it's hard to use the side rangefinder to determine when we've
crossed ramparts, we are reverting to the ramparts autonomous from before
we changed it to use the side rangefinder. Moat and rock wall, which used
to use the ramparts autonomous, have now been changed to run a new moat
autonomous, which is the version of the ramparts autonomous that uses the
side rangefinder.